edulink-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edulink-agent"
version = "0.1.0"
description = "Playwright-powered Edulink automation service that surfaces homework, behaviour, and mailbox summaries."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "fastapi>=0.110.0",
  "playwright>=1.45.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "python-dateutil>=2.9.0",
  "uvicorn>=0.29.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "ruff>=0.5.0"
]

[project.scripts]
edulink-agent = "edulink_agent.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
